+++ bukkit/src/main/java/io/github/silthus/template/TemplateBukkitPlugin.java
@@ -1,0 +1,52 @@
+/*
+ * This file is part of <REPLACE_WITH_PLUGIN>, licensed under the MIT License.
+ * Copyright (c) <REPLACE_ME> <REPLACE_ME@REPLACE_ME>
+ * Copyright (c) contributors
+ *
+ *  Permission is hereby granted, free of charge, to any person obtaining a copy
+ *  of this software and associated documentation files (the "Software"), to deal
+ *  in the Software without restriction, including without limitation the rights
+ *  to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ *  copies of the Software, and to permit persons to whom the Software is
+ *  furnished to do so, subject to the following conditions:
+ *
+ *  The above copyright notice and this permission notice shall be included in all
+ *  copies or substantial portions of the Software.
+ *
+ *  THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ *  IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ *  FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ *  AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ *  LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ *  OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+ *  SOFTWARE.
+ */
+package io.github.silthus.template;
+
+import io.github.silthus.template.listener.PlayerListener;
+import java.io.File;
+import org.bukkit.plugin.PluginDescriptionFile;
+import org.bukkit.plugin.java.JavaPlugin;
+import org.bukkit.plugin.java.JavaPluginLoader;
+import org.jetbrains.annotations.NotNull;
+
+public class TemplateBukkitPlugin extends JavaPlugin {
+
+    public TemplateBukkitPlugin() {
+    }
+
+    // required for testing
+    public TemplateBukkitPlugin(@NotNull JavaPluginLoader loader, @NotNull PluginDescriptionFile description, @NotNull File dataFolder, @NotNull File file) {
+        super(loader, description, dataFolder, file);
+    }
+
+    @Override
+    public void onEnable() {
+        getServer().getPluginManager().registerEvents(new PlayerListener(), this);
+    }
+
+    @Override
+    public void onDisable() {
+
+    }
+}