+++ bukkit/src/test/java/io/github/silthus/template/BukkitTests.java
@@ -1,0 +1,54 @@
+/*
+ * This file is part of <REPLACE_WITH_PLUGIN>, licensed under the MIT License.
+ * Copyright (c) <REPLACE_ME> <REPLACE_ME@REPLACE_ME>
+ * Copyright (c) contributors
+ *
+ *  Permission is hereby granted, free of charge, to any person obtaining a copy
+ *  of this software and associated documentation files (the "Software"), to deal
+ *  in the Software without restriction, including without limitation the rights
+ *  to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ *  copies of the Software, and to permit persons to whom the Software is
+ *  furnished to do so, subject to the following conditions:
+ *
+ *  The above copyright notice and this permission notice shall be included in all
+ *  copies or substantial portions of the Software.
+ *
+ *  THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ *  IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ *  FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ *  AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ *  LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ *  OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+ *  SOFTWARE.
+ */
+package io.github.silthus.template;
+
+import be.seeseemelk.mockbukkit.MockBukkit;
+import be.seeseemelk.mockbukkit.MockPlugin;
+import be.seeseemelk.mockbukkit.ServerMock;
+import org.junit.jupiter.api.AfterAll;
+import org.junit.jupiter.api.BeforeAll;
+import org.junit.jupiter.api.BeforeEach;
+
+import static org.assertj.core.api.Assertions.assertThat;
+
+public abstract class BukkitTests {
+    protected static ServerMock server;
+
+    protected TemplateBukkitPlugin plugin;
+
+    @BeforeAll
+    static void beforeAll() {
+        server = MockBukkit.mock(new CustomServerMock());
+    }
+
+    @BeforeEach
+    void setUp() {
+        plugin = MockBukkit.load(TemplateBukkitPlugin.class);
+    }
+
+    @AfterAll
+    static void afterAll() {
+        MockBukkit.unmock();
+    }
+}